Restructured bulk module
