+++ src/app/admin/admin.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 
 import { SharedModule } from '../shared/shared.module';
+import { BulkModule } from '../bulk/bulk.module';
 import { AdminRoutingModule } from './routing/admin-routing.module';
 
 import { DashboardComponent,
@@ -19,7 +20,8 @@ import { BulkComponent } from './components/bulk/bulk.component';
   ],
   imports: [
     SharedModule,
-    AdminRoutingModule
+    AdminRoutingModule,
+    BulkModule
   ]
 })
 export class AdminModule { }
