+++ src/app/bulk/bulk.module.ts
@@ -6,24 +6,30 @@ import { SharedModule } from '../shared/shared.module';
 import { BulkRoutingModule } from './routing/bulk-routing.module';
 import { effects, reducer } from './store';
 
-import { BulkDetailsComponent, BulkUploadComponent, BulkSummaryComponent } from './components';
+import {
+  BulkDetailsComponent, BulkUploadComponent, BulkSummaryComponent,
+  BulkSummaryTableComponent, BulkSummaryQuestionComponent
+} from './components';
 
 
 @NgModule({
   declarations: [
     BulkDetailsComponent,
     BulkUploadComponent,
-    BulkSummaryComponent
+    BulkSummaryComponent,
+    BulkSummaryTableComponent,
+    BulkSummaryQuestionComponent
   ],
   imports: [
     SharedModule,
     BulkRoutingModule,
 
-     //ngrx feature store
-     StoreModule.forFeature('bulk', reducer),
+    //ngrx feature store
+    StoreModule.forFeature('bulk', reducer),
 
     //  //ngrx effects
-     EffectsModule.forFeature(effects),
-  ]
+    EffectsModule.forFeature(effects),
+  ],
+  exports: [BulkSummaryTableComponent, BulkSummaryQuestionComponent]
 })
 export class BulkModule { }
