+++ src/app/bulk/components/bulk-summary/bulk-summary-question/bulk-summary-question.component.html
