+++ src/app/bulk/components/bulk-summary/bulk-summary-question/bulk-summary-question.component.ts
@@ -2,12 +2,12 @@ import { Component, Input, OnInit, OnChanges, OnDestroy } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { Store } from '@ngrx/store';
-import { AppStore } from '../../../core/store/app-store';
-import { BulkUploadFileInfo, Question, Category } from '../../../model';
-import { QuestionActions } from '../../../core/store/actions';
+import { AppStore } from '../../../../core/store/app-store';
+import { BulkUploadFileInfo, Question, Category } from '../../../../model';
+import { QuestionActions } from '../../../../core/store/actions';
 import { Subscription } from 'rxjs/Subscription';
 import { MatTableDataSource } from '@angular/material';
-import { Utils } from '../../../core/services';
+import { Utils } from '../../../../core/services';
 import { AngularFireStorage } from 'angularfire2/storage';
