+++ src/app/bulk/components/bulk-summary/bulk-summary-table/bulk-summary-table.component.html
