+++ src/app/bulk/components/index.ts
@@ -1,14 +1,20 @@
 import { BulkDetailsComponent } from './bulk-details/bulk-details.component';
 import { BulkUploadComponent } from './bulk-upload/bulk-upload.component';
 import { BulkSummaryComponent } from './bulk-summary/bulk-summary.component';
+import { BulkSummaryTableComponent } from './bulk-summary/bulk-summary-table/bulk-summary-table.component';
+import { BulkSummaryQuestionComponent } from './bulk-summary/bulk-summary-question/bulk-summary-question.component';
 export {
   BulkDetailsComponent,
   BulkUploadComponent,
-  BulkSummaryComponent
+  BulkSummaryComponent,
+  BulkSummaryTableComponent,
+  BulkSummaryQuestionComponent
 };
 
 export default [
   BulkDetailsComponent,
   BulkUploadComponent,
-  BulkSummaryComponent
+  BulkSummaryComponent,
+  BulkSummaryTableComponent,
+  BulkSummaryQuestionComponent,
 ];
