+++ src/app/shared/components/index.ts
@@ -2,16 +2,12 @@ import { QuestionComponent } from './question/question.component';
 import { QuestionsComponent } from './questions/questions.component';
 import { QuestionsSearchComponent } from './questions-search/questions-search.component';
 import { QuestionsTableComponent } from './questions-table/questions-table.component';
-import { BulkSummaryTableComponent } from './bulk-summary-table/bulk-summary-table.component';
-import { BulkSummaryQuestionComponent } from './bulk-summary-question/bulk-summary-question.component';
 import { QuestionFormComponent } from './question-form/question-form.component';
 export {
     QuestionComponent,
     QuestionsComponent,
     QuestionsSearchComponent,
     QuestionsTableComponent,
-    BulkSummaryTableComponent,
-    BulkSummaryQuestionComponent,
     QuestionFormComponent
 };
 
@@ -20,7 +16,5 @@ export default [
     QuestionsComponent,
     QuestionsSearchComponent,
     QuestionsTableComponent,
-    BulkSummaryTableComponent,
-    BulkSummaryQuestionComponent,
     QuestionFormComponent
 ];
